rpi_code_borohexa/requirements.txt
pymavlink
mavproxy
numpy
opencv-python
pyserial
matplotlib
lxml
